+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `ETV_SLOW_API_MS` - milliseconds threshold for logging slow API calls (at DEBUG level)
     - This is currently limited to *Jellyfin*
   - `ETV_JF_PAGE_SIZE` - page size for library scan API calls to Jellyfin; default value is 10
+- Add `Select All` button to media pages by @Erotemic
 
 ### Fixed
 - Fix startup on systems unsupported by NvEncSharp